Revert "Pass manager refactoring: add passmanager module (#10124)"
This reverts commit fbd64d95d9002315819c5c59d9cfd1d53de5bddb. The follow
on PR to this one #10127 which is making internal refactors to the pass
manager code now that's it a standalone module is still undergoing
active review and the scope of the PR is sufficiently large that it
likely won't be viable for the pending 0.25.0 release. This commit
temporarily reverts #10124 which was the first step of creating a module
by porting the pass manager code in it's current form to a standalone
module so that we're not committed to the API as part of the 0.25.0
release to give more time for #10127 to finalize what the eventual
`qiskit.passmanager` API will look like.

This revert should itself be reverted after 0.25.0rc1 is tagged and
the main branch opens up for 0.45.0 development. As this revert is
not an indication that we did not want #10124 it's just to avoid
committing to the API prematurely.
