--- docs/apidocs/passmanager.rst
@@ -1,6 +1,0 @@
-.. _qiskit-passmanager:
-
-.. automodule:: qiskit.passmanager
-   :no-members:
-   :no-inherited-members:
-   :no-special-members:
