+++ qiskit/transpiler/basepasses.py
@@ -13,18 +13,61 @@
 """Base transpiler passes."""
 
 from abc import abstractmethod
-
-from qiskit.passmanager.base_pass import GenericPass
-from qiskit.passmanager.propertyset import PropertySet
-
+from collections.abc import Hashable
+from inspect import signature
+from .propertyset import PropertySet
 from .layout import TranspileLayout
 
 
-class BasePass(GenericPass):
+class MetaPass(type):
+    """Metaclass for transpiler passes.
+
+    Enforces the creation of some fields in the pass while allowing passes to
+    override ``__init__``.
+    """
+
+    def __call__(cls, *args, **kwargs):
+        pass_instance = type.__call__(cls, *args, **kwargs)
+        pass_instance._hash = hash(MetaPass._freeze_init_parameters(cls, args, kwargs))
+        return pass_instance
+
+    @staticmethod
+    def _freeze_init_parameters(class_, args, kwargs):
+        self_guard = object()
+        init_signature = signature(class_.__init__)
+        bound_signature = init_signature.bind(self_guard, *args, **kwargs)
+        arguments = [("class_.__name__", class_.__name__)]
+        for name, value in bound_signature.arguments.items():
+            if value == self_guard:
+                continue
+            if isinstance(value, Hashable):
+                arguments.append((name, type(value), value))
+            else:
+                arguments.append((name, type(value), repr(value)))
+        return frozenset(arguments)
+
+
+class BasePass(metaclass=MetaPass):
     """Base class for transpiler passes."""
 
+    def __init__(self):
+        self.requires = []  # List of passes that requires
+        self.preserves = []  # List of passes that preserves
+        self.property_set = PropertySet()  # This pass's pointer to the pass manager's property set.
+        self._hash = hash(None)
+
+    def __hash__(self):
+        return self._hash
+
+    def __eq__(self, other):
+        return hash(self) == hash(other)
+
+    def name(self):
+        """Return the name of the pass."""
+        return self.__class__.__name__
+
     @abstractmethod
-    def run(self, dag):  # pylint: disable=arguments-differ
+    def run(self, dag):
         """Run a pass on the DAGCircuit. This is implemented by the pass developer.
 
         Args:
